code organisation and cleanup
