+++ cli/static/run_node.js
@@ -119,7 +119,7 @@ async function main(document, window) {
   document.getElementById("start").addEventListener("click", async () => {
     document.querySelectorAll('.config').forEach(elem => elem.disabled = true);
 
-    await window.driver.start(window.config)
+    await window.driver.start(window.config);
     document.getElementById("peer-id").innerText = await window.driver.local_peer_id();
     document.querySelectorAll(".status").forEach(elem => elem.style.visibility = "visible");
   });
