+++ node-wasm/src/lib.rs
@@ -9,4 +9,4 @@ mod worker;
 mod wrapper;
 
 /// Alias for a `Result` with the error type [`JsError`].
-pub type Result<T> = std::result::Result<T, JsError>;
+pub type Result<T, E = JsError> = std::result::Result<T, E>;
